Health and HealthBars
added Health Component
added HealthBar and pf

+++ Assets/Scripts/Grid/CityBuilding/IBuilding.cs
@@ -1,8 +1,9 @@
 using UnityEngine;
 
+
 public interface IBuilding
 {
     void Init() { }
     bool CanPlace(Vector3 worldPos) => true;
-    void TakeDamage(int damage){}
+    void TakeDamage(float damage){}
 }
